Remove compile warnings for building for 64 bit

+++ src/Blur/UIImage+MRImageEffects.m
@@ -154,7 +154,7 @@ vImage_Buffer vImageBuffer_InitWithCGContext(CGContextRef contextRef) {
             // ... if d is odd, use three box-blurs of size 'd', centered on the output pixel.
             //
             CGFloat inputRadius = blurRadius * 1;
-            NSUInteger radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
+            uint32_t radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
             if (radius % 2 != 1) {
                 radius += 1; // force radius to be odd so that the three box-blur methodology works.
             }
